+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- refactored server connection into a single provider
+
 ## [v0.5.4] - 2026-07-08
 
 ### Added